requirements.txt
python-dotenv
chromadb
openai